requirements.txt
pygame==2.5.2
openai==0.28.1
python-dotenv==1.0.0
numpy==1.26.0